Rename/Video.jsx파일을 components/ResultPage폴더로 옮김
